# packages/api-contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Jomboy Lavka API
  version: 1.0.0
  description: Unified API for Jomboy Lavka client applications
servers:
  - url: https://api.jomboy-lavka.uz/api/v1
    description: Production
  - url: http://localhost:4010/api/v1
    description: Mock server
security:
  - bearerAuth: []
paths:
...
  /orders:
    get:
      tags: [Order]
      summary: List customer orders
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Order list
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: './schemas/common.yaml#/components/schemas/Order'
    post:
      tags: [Order]
      summary: Create order
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [darkstore_id, items, delivery_address, payment_method]
              properties:
                darkstore_id:
                  type: string
                  format: uuid
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      product_id:
                        type: string
                        format: uuid
                      quantity:
                        type: number
                delivery_address:
                  $ref: './schemas/common.yaml#/components/schemas/DeliveryAddress'
                payment_method:
                  type: string
                  enum: [payme, click]
                comment:
                  type: string
                  maxLength: 500
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                type: object
                properties:
                  order_id:
                    type: string
                    format: uuid
                  status:
                    type: string
                  total_amount:
                    type: integer
                  delivery_fee:
                    type: integer
                  estimated_delivery_minutes:
                    type: integer
  /orders/{order_id}:
    get:
      tags: [Order]
      summary: Get order status
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: './schemas/common.yaml#/components/schemas/Order'
  /orders/{order_id}/cancel:
    post:
      tags: [Order]
      summary: Cancel order (before ASSEMBLY)
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Cancelled
...
  /tickets:
    get:
      tags: [Support]
      summary: Ticket queue (operator)
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Tickets
          content:
            application/json:
              schema:
                type: object
                properties:
                  tickets:
                    type: array
                    items:
                      $ref: './schemas/common.yaml#/components/schemas/Ticket'
    post:
      tags: [Support]
      summary: Create ticket
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                order_id:
                  type: string
                type:
                  type: string
                description:
                  type: string
                priority:
                  type: string
  /tickets/{ticket_id}:
    get:
      tags: [Support]
      summary: Get ticket details
      parameters:
        - name: ticket_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ticket
          content:
            application/json:
              schema:
                $ref: './schemas/common.yaml#/components/schemas/Ticket'
  /tickets/{ticket_id}/refund-decision:
    post:
      tags: [Support]
      summary: Refund decision
      parameters:
        - name: ticket_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision:
                  type: string
                  enum: [approve, reject, partial]
                amount:
                  type: integer
                comment:
                  type: string
      responses:
        '200':
          description: Decision recorded
...
  /support/ai/suggest:
    post:
      tags: [Support, Phase3, AI]
      summary: Mock RAG answer suggestions
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                question:
                  type: string
                ticket_type:
                  type: string
      responses:
        '200':
          description: Suggestions with confidence
...
  /tickets/{ticket_id}/auto-refund-eligibility:
    get:
      tags: [Support, Phase3]
      summary: Auto-refund rules check
      parameters:
        - name: ticket_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Eligibility result
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    OrderId:
      name: order_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: X-Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid